+++ online/public/bhall/basic/index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <link rel="shortcut icon" href="../../favicon.ico" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta name="theme-color" content="#000000" />
+    <meta
+      name="description"
+      content="Zometool models designed by Brian Hall"
+    />
+    <title>Brian Hall's Zometool Models</title>
+    <link rel="stylesheet" href="../../App.css">
+    <link rel="manifest" href="./manifest.json">
+  </head>
+  <body>
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="root"></div>
+    <script type="module" src="../../../modules/bhall-basic.js"></script>
+  </body>
+</html>
