+++ online/public/browser/index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <link rel="shortcut icon" href="../favicon.ico" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta name="theme-color" content="#000000" />
+    <meta
+      name="description"
+      content="Online browser for vZome (https://vzome.com) files in a vzome-sharing GitHub repository"
+    />
+    <title>vZome GitHub Browser</title>
+    <link rel="stylesheet" href="../App.css">
+    <link rel="manifest" href="./manifest.json">
+  </head>
+  <body>
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="root"></div>
+    <script type="module" src="../../modules/vzome-browser.js"></script>
+  </body>
+</html>
